+++ Makefile
@@ -9,9 +9,12 @@ else
 endif
 
 ifeq (add,$(firstword $(MAKECMDGOALS)))
-  # use the rest as arguments for "run"
   RUN_ARGS := $(wordlist 2,$(words $(MAKECMDGOALS)),$(MAKECMDGOALS))
-  # ...and turn them into do-nothing targets
+  $(eval $(RUN_ARGS):;@:)
+endif
+
+ifeq (remove,$(firstword $(MAKECMDGOALS)))
+  RUN_ARGS := $(wordlist 2,$(words $(MAKECMDGOALS)),$(MAKECMDGOALS))
   $(eval $(RUN_ARGS):;@:)
 endif
 
@@ -22,13 +25,25 @@ run:
 	$(PYTHON) -m bot
 
 install-beautifier:
-	pip install black isort
+	@pip install black isort
+	@echo "Successfully installed beautifier!"
 
 beautify:
-	black .
-	isort .
+	@black .
+	@echo "Successfully beautified code!"
+	@isort .
+	@echo "Successfully sorted imports!"
 
 add:
 	@echo "Adding new module..."
-	poetry add $(RUN_ARGS)
-	poetry export -f requirements.txt --output requirements.txt --without-hashes
+	@poetry add $(RUN_ARGS)
+	@echo "Updating requirements.txt..."
+	@poetry export -f requirements.txt --output requirements.txt --without-hashes
+	@echo "Successfully added new module!"
+
+remove:
+	@echo "Removing module..."
+	@poetry remove $(RUN_ARGS)
+	@echo "Updating requirements.txt..."
+	@poetry export -f requirements.txt --output requirements.txt --without-hashes
+	@echo "Successfully removed module!"
