+++ bot/commands/levels/__assets/idle.png
