+++ bot/commands/levels/__assets/offline.png
