+++ bot/commands/levels/__assets/streaming.png
