+++ bot/commands/levels/__box_generator.py
